Move the (as requested) all enddpoints from AirTrafficEndpoint to domain based endpoints
